requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
et-xmlfile==1.1.0
frozenlist==1.4.0
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
meraki==1.36.0
multidict==6.0.4
numpy==1.25.2
openpyxl==3.1.2
pandas==2.1.0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rich==13.5.3
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
